requirements.txt
telethon==1.37.0
customtkinter==5.2.2
openpyxl==3.1.5
fpdf2==2.8.2
jinja2==3.1.4
Pillow==11.0.0
cryptg==0.4.0
